build(core): use the same fake email across tables when obfuscating data
This makes associating related data easier in the staging environment e.g. it
is easier to see which emails were sent by the client vs the paralegal.

We also fake email addresses in submission data instead of redacting them to
make testing associating submissions with issues easier. This could be removed
later but probably doesn't need to be.