build(core): only obfuscate fields with data
Previously we would generate fake staging environment data for a bunch of model
fields irrespective of whether they actually had a value in the production
environment. This caused an issue for our simple method of using the same email
across different models as the same email would be used for all email model
fields with empty values. We just add some guard conditions to prevent this.

This also gives us staging data that is more similar to the production data i.e.
empty fields in production are also empty in staging.