fix(submission): correct spelling of keys used to get specific answers
The 2 keys - which correspond to 2 questions in the intake form - changed are:

- LEGAL_CENTRE_REFERRER
- AVAILABILITY

The former is correct in the intake form so this value was not being copied from
the submission to the issue correctly. The latter was incorrect in Clerk and the
intake form so it needs to be changed in both places.

Also update tests that use these values.