ops: adjust how we deploy development branches to the staging environment
As part of running the github test action docker images are built to run the
tests and, if the tests are successful, pushed to docker hub for later
deployment. Previous to this change docker builds of non-prod & non-staging
branches would be tagged as "dev" and could then be deployed to the staging
environment.

This was largely fine but the naming was confusing and restoring the staging
environment from production didn't function correctly as we always used the
*staging* docker compose file when running the staging restore script -
specifically migrations would not work if the migrations files were not also
included in the docker image which would be the case for any migration file
in the "dev" docker image.

Rather than mess around trying to fix up which docker compose file we use I just
simplified a bit and use the staging tag for any non-prod docker build.

For some background, see commit ad5231d04d9eafc655aacaac083a8902446e0783.