build: update wagtail CMS dependency
As we are about to make some updates to the CMS we might as well be using the
latest version.

Note that we have to make a change so that CMS blog search tests pass. The code that indexes Wagtail pages is now triggered from a Django on_commit callback so we have to make sure the callback is executed when saving the page.