test(web): fix failing test when factory document file ext is html or pdf
The test_tracked_document_download and test_untracked_document_download tests
fail if the factory-generated document file extension is "html". To fix, we just
set the document file type explicitly to "pdf" and remove the content
disposition check as it would fail - it depends on the file type and "inline" is
used for pdfs - but actually isn't a useful test to warrant keeping & fixing.