fix(web): prevent error accessing the request context when getting banner info
In brief, an error occurs when the banner method is called without a request
context. The originating cause is not clear - it looks like the db connection
has been lost - but guarding against it seems like a reasonable course of
action. See
https://anika-legal.sentry.io/issues/7049868788/events/38d322325fbb4129872713fa0085035a/.

Also refactor to reduce the number of return statements.