feat: display custom 403 & 404 error pages in clerk frontend
Display custom error pages in the clerk frontend if we cannot access (403) or
cannot find (404) a resource when routing or attempting to render a view.
Previously pages that were not found would display the website 404 page and
forbidden pages would return a JSON error response.

In addition to improving the UX this also resolves TEC-1708 - a type error
ultimately resulting from a user attempting to access a case for which they do
not have the required permissions. To address this we check that the case exists
& can be accessed at the Django routing and view rendering layer and add some
fallback error handling to the rendered react page.

See https://linear.app/anika-legal/issue/TEC-1708