fix: adjust / improve error handling for frontend email list & thread display
The API call that retrieves lists of emails by thread (or a list with a single
item for a particular thread) 404's if there are no results i.e. no email
threads associated with an issue or no thread with a specific slug. I adjusted
this so that we return an empty list instead if there are no email threads
associated with an issue. This makes the error handling on the client side a bit
tidier. This is fine for now but we should look at improving the API down the
track as it is a bit crufty, see TEC-1717.

Also added frontend error handling to explicitly handle the 404 error when there
is no thread with a specific slug and we throw on other errors so that they are
handled at our error boundary.

Resolves TEC-1715 & resolves TEC-1716.