fix(microsoft): don't log an error AND raise exception when msgraph API call fails
If we do both then we get an issue raised in our error tracking service even if
the calling code handles the raised exception. We should either let the
exception be handled and possibly logged by the calling code or let it fall
through to be handled by our error tracking service.