fix(microsoft): KeyError: 'grantedToV2' when listing sharepoint permissions
When fixing the previous error here (see commit
59e7e308083c8aa382fc35c0b129495cc82319c1) I misunderstood the structure of the
msgraph API permission resource type. Each permission contains a so-called facet
that indicates what type of permission it is. In our code we are expecting only
a specific facet, "grantedToV2", which is not always the case and results in the
KeyError.

To correct this we return all permission resource types from the msgraph API
call and check for the specific facet we need in the calling code. This is also
more flexible if we want to use other permission facets later. Resolves
TEC-1727.

Also update tests because of the changed return type of the msgraph permissions
API call.