fix(microsoft): add check for group owners when testing for msgraph permissions
This is a minor issue that I have fixed mostly because it causes confusion.
Sharepoint group owners - who are exclusively Anika admin users including myself
- don't show up as having full permissions for sharepoint folders in the Clerk
interface because we only test for regular members. This changes corrects that
issue.

Note that the attribute we assign to indicate full sharepoint folder access,
has_coordinator_perms, is a bit of a misnomer as highlighted by this change; you
can have full sharepoint folder access but not be a coordinator. I have left
this for now but might be worth changing down the track.