fix(frontend): logException after deleting draft email
The (partial) stack trace is:

  Error: Object captured as exception with keys: error, status
      at logException (../../src/comps/error-boundary.tsx:24:5)
      at ErrorBoundary.componentDidCatch (../../src/comps/error-boundary.tsx:41:5)

See https://anika-legal.sentry.io/issues/7097268022/ for full details.

This occurs because we get a render after we delete a draft email but before
redirecting to the email list page. The email query fails (as it was deleted)
and we throw the resulting (404) error. The error is captured at the error
boundary and we try to send it to our error tracking service (Sentry). Their
software throws an error because they (rightly) do not expect that kind of
error.

We workaround this by returning before we check for email query errors if we are
deleting or have deleted the email. Resolves TEC-1728.

We should refactor this to remove the unnecessary query for the deleted email.
See TEC-1729.